Smoke tests for OSP and RHV
